requirements.txt
--index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://pypi.org/simple

fastapi==0.115.4
uvicorn[standard]
firebase-admin
python-dotenv==1.0.1
pydantic==2.9.2
pandas==2.2.3
numpy
scikit-learn
plotly==5.24.1
matplotlib
google-genai
httpx==0.27.2

# CPU-only PyTorch + friends (for Python 3.13)
torch==2.9.1+cpu
pytorch-lightning==2.6.0
pytorch-forecasting==1.5.0
